fix front end from p'yu comments
