+++ apps/cugetreg/src/routes/+page.ts
@@ -1,0 +1,5 @@
+import { redirect } from '@sveltejs/kit'
+
+export const load = () => {
+  throw redirect(307, '/course')
+}
